+++ app/src/main/java/com/apptive/braini/_enums/Side.kt
@@ -1,0 +1,5 @@
+package com.apptive.braini._enums
+
+enum class Side {
+    LEFT, RIGHT
+}
