--- app/src/main/java/com/example/braini/_enums/Side.kt
@@ -1,5 +1,0 @@
-package com.example.braini._enums
-
-enum class Side {
-    LEFT, RIGHT
-}
