--- app/src/main/java/com/example/braini/domain/extension.kt
@@ -1,7 +1,0 @@
-package com.example.braini.domain
-
-import android.util.Log
-
-fun String.log(header: String = "브레이니") {
-    Log.d(header, this)
-}
