--- app/src/main/java/com/example/braini/domain/typealias.kt
@@ -1,8 +1,0 @@
-package com.example.braini.domain
-
-import androidx.activity.compose.ManagedActivityResultLauncher
-import androidx.activity.result.ActivityResult
-import androidx.activity.result.IntentSenderRequest
-
-/** Activity Result API: Intent Sender Request */
-typealias IntentSenderLauncher = ManagedActivityResultLauncher<IntentSenderRequest, ActivityResult>
